Create typing stub
